refactor: use browser modules
Makes for smaller bundle becuase we can use `is-buffer` instead of including `Buffer` in the bundle.

resolves #8
resolves #9
