+++ src/cmd/client/import.js
@@ -1,11 +1,7 @@
-/* eslint-env browser */
-
 const toUri = require('multiaddr-to-uri')
-// const toStream = require('async-iterator-to-stream')
-const FormData = require('form-data')
 const CID = require('cids')
 const { ok } = require('../../lib/fetch')
-const { isBrowser, isNode } = require('../../lib/env')
+const { toFormData } = require('../../lib/form-data')
 
 module.exports = (fetch, config) => {
   return async (input, options) => {
@@ -21,42 +17,4 @@ module.exports = (fetch, config) => {
     const data = await res.json()
     return new CID(data['/'])
   }
-}
-
-async function toFormData (input) {
-  // If not an async iterator, this must conform to whatever FormData allows
-  if (!input[Symbol.asyncIterator]) {
-    const formData = new FormData()
-    formData.append('file', input)
-    return formData
-  }
-
-  // In Node.js, FormData can be passed a stream so no need to buffer
-  if (isNode) {
-    const formData = new FormData()
-    const bufs = []
-    for await (const chunk of input) {
-      bufs.push(Buffer.from(chunk))
-    }
-    // FIXME: the below does not work! It should do, but only the first chunk
-    // gets uploaded :(
-    // formData.append('file', toStream(input))
-    formData.append('file', Buffer.concat(bufs))
-    return formData
-  }
-
-  // In the browser there's _currently_ no streaming upload, buffer up our
-  // async iterator chunks and append a big Blob :(
-  if (isBrowser) {
-    // One day, this will be browser streams
-    const formData = new FormData()
-    const bufs = []
-    for await (const chunk of input) {
-      bufs.push(Buffer.isBuffer(chunk) ? chunk.buffer : chunk)
-    }
-    formData.append('file', new Blob(bufs))
-    return formData
-  }
-
-  throw new Error('unknown environment')
 }
